+++ react/modules/interpolateLink.ts
@@ -1,3 +1,5 @@
+import { searchSlugify } from '@vtex/slugify'
+
 import { getMappingFn, AvailableContext } from './mappings'
 
 interface Params {
@@ -20,7 +22,9 @@ export default function interpolateLink(params: Params) {
       `{${namespace ? `${namespace}.${key}` : key}}`,
       'g'
     )
-    resolvedLink = resolvedLink.replace(regex, variables[key])
+
+    // we slugify the context value so we don't end up with invalid characters on the URL
+    resolvedLink = resolvedLink.replace(regex, searchSlugify(variables[key]))
   }
   // Replace not found variables with empty string
   if (namespace) {