+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- Use search API slugify method for product brand.
+
 ## [0.7.0] - 2020-09-10
 ### Added
 - `scrollTo` option to perform scroll after navigation.